+++ lib/petfinder/mash.rb
@@ -1,0 +1,5 @@
+module Petfinder
+  class Mash < Hashie::Mash
+    disable_warnings
+  end
+end