+++ spec/client_spec.rb
@@ -1,62 +1,62 @@
-require 'spec_helper'
+RSpec.describe Petfinder::Client do
 
-describe Petfinder::Client do
   before do
-    @client = Petfinder::Client.new('1234567', 'mysekrit')
+    stub_post("https://api.petfinder.com/v2/oauth2/token", "token.json")
   end
 
-  it "should get a specific pet by id" do
-    stub_get("http://api.petfinder.com/pet.get?key=1234567&id=123", "pet.xml")
-    pet = @client.pet(123)
-    pet.name.should == 'Charlie'
+  let(:client) { Petfinder::Client.new('api_key', 'api_secret') }
+
+  it 'returns array of animals with pagination' do
+    stub_get("https://api.petfinder.com/v2/animals", "animals.json")
+    animals, pagination = client.animals
+
+    expect(animals.first.name).to eq("Clarissa")
+    expect(pagination.current_page).to eq(1)
   end
 
-  it "should get a random pet" do
-    stub_get("http://api.petfinder.com/pet.getRandom?key=1234567&output=full", "pet.xml")
-    pet = @client.random_pet
-    pet.name.should == 'Charlie'
+  it 'returns a specific animal' do
+    stub_get("https://api.petfinder.com/v2/animals/123", "animal.json")
+    animal = client.animal(123)
+
+    expect(animal.name).to eq("Apollo")
   end
 
-  it "should find pets matching a query" do
-    stub_get("http://api.petfinder.com/pet.find?key=1234567&animal=dog&location=77007", "pet_list.xml")
-    pets = @client.find_pets('dog', '77007')
-    pets.first.name.should == 'Petey'
+  it 'returns array of animal types' do
+    stub_get("https://api.petfinder.com/v2/types", "types.json")
+    types = client.types
+
+    expect(types.length).to eq(8)
+    expect(types[1].name).to eq("Cat")
   end
 
-  it "should get the list of breeds for the given animal type" do
-    stub_get("http://api.petfinder.com/breed.list?key=1234567&animal=horse", "breed_list.xml")
-    breeds = @client.breeds('horse')
-    breeds.first.should == "Appaloosa"
+  it 'returns a specific animal type' do
+    stub_get("https://api.petfinder.com/v2/types/dog", "type.json")
+    type = client.type('dog')
+
+    expect(type.name).to eq("Dog")
   end
 
-  it "should get a specific shelter by id" do
-    stub_get("http://api.petfinder.com/shelter.get?key=1234567&id=123", "shelter.xml")
-    shelter = @client.shelter(123)
-    shelter.name.should == 'Lucky Dog Rescue'
+  it 'returns list of breeds for animal type' do
+    stub_get("https://api.petfinder.com/v2/types/bird/breeds", "breeds.json")
+    breeds = client.breeds('bird')
+
+    expect(breeds.length).to eq(39)
+    expect(breeds.first.name).to eq("African Grey")
   end
 
-  it "should find shelters matching a query" do
-    stub_get("http://api.petfinder.com/shelter.find?key=1234567&location=77007", "shelter_list.xml")
-    shelters = @client.find_shelters('77007')
-    shelters.first.name.should == 'Pup Squad Animal Rescue'
+  it 'returns array of organizations with pagination' do
+    stub_get("https://api.petfinder.com/v2/organizations", "organizations.json")
+    organizations, pagination = client.organizations
+
+    expect(organizations.first.name).to eq("Alberta Bulldog Rescue - Red Deer")
+    expect(pagination.current_page).to eq(1)
   end
 
-  it "should find shelters matching a breed" do
-    stub_get("http://api.petfinder.com/shelter.listByBreed?key=1234567&animal=horse&breed=arabian", "shelter_list.xml")
-    shelters = @client.find_shelters_by_breed('horse', 'arabian')
-    shelters.first.name.should == 'Pup Squad Animal Rescue'
-  end
+  it 'returns a specific organization' do
+    stub_get("https://api.petfinder.com/v2/organizations/mo201", "organization.json")
+    organization = client.organization('mo201')
 
-  it "should get the list of pets available for the given shelter" do
-    stub_get("http://api.petfinder.com/shelter.getPets?key=1234567&id=123", "pet_list.xml")
-    pets = @client.shelter_pets(123)
-    pets.first.name.should == 'Petey'
-  end
-
-  it "should get an authorization token" do
-    digest = @client.send(:digest_key_and_secret)
-    stub_get("http://api.petfinder.com/auth.getToken?key=1234567&sig=#{digest}", "auth.xml")
-    @client.token.should == '39038a000f404335d90bab019c0c7e2d'
+    expect(organization.name).to eq("Animal Shelter Of Linn County")
   end
 
 end